+++ services/ui-src/src/components/TriageExternalLandingPage.tsx
@@ -1,11 +1,19 @@
 import React from "react";
-import { TriageLandingPageConfig } from "../../libs/landingPageLib";
 import { Button } from "@cmsgov/design-system";
-import PageTitleBar from "../../components/PageTitleBar";
+import PageTitleBar from "./PageTitleBar";
+
+export interface TriageLandingPageConfig {
+  pageTitle: string;
+  // logo passed as <img /> element
+  logoJSX: JSX.Element;
+  // TriageExternalLandingPage component handles positioning, but use
+  // <p> tags to handle spacing in the descriptionJSX element
+  descriptionJSX: JSX.Element;
+  buttonLabel: string;
+}
 
 /** Config-driven template to build landing pages that link to another system.
- * See {@link TriageLandingPageConfig} for details on the config object. All
- * configs are maintained in `libs/landingPageLib.ts`.
+ * See {@link TriageLandingPageConfig} for details on the config object.
  */
 const TriageExternalLandingPage = ({
   pageTitle,
