--- services/ui-src/src/page/landing/LandingPages.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-
-export const MacPROLogo = <img src="/assets/images/" alt="MacPRO Logo" />;
-export const MMDLLogo = <img src="/assets/images/" alt="MMDL Logo" />;
-
-export const ABPLandingPageContent = <></>;
-export const MedicaidEligibilityLandingPageContent = <></>;
-export const CHIPEligibilityLandingPageContent = <></>;
